set up react router
