+++ package.json
@@ -43,6 +43,7 @@
 		]
 	},
 	"devDependencies": {
-		"gh-pages": "^5.0.0"
+		"gh-pages": "^5.0.0",
+		"react-router-dom": "^6.14.1"
 	}
 }
