--- src/App.css
@@ -1,123 +1,0 @@
-.App {
-  overflow: hidden;
-  /* Hide scrollbars */
-  height: 100vh;
-  width: 100vw;
-  /* Initialize flexbox */
-  display: flex;
-  flex-direction: column;
-  font-family: 'Roboto Slab', serif;
-  background-size: cover;
-
-}
-
-.nightMode {
-  color: white;
-  background-color: black;
-  background-image: url('./assets/backgroundUniverseDark.svg');
-  transition: all 1s;
-}
-
-.dayMode {
-  color: #000000;
-  background-color: white;
-  background-image: url('./assets/backgroundGradientLight.png');
-  transition: all 0.1s;
-
-}
-
-.NavBar {
-  overflow: hidden;
-  top: 0;
-  position: fixed;
-}
-
-#TopNav {
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: flex-start;
-}
-
-#tassel {
-  height: 20vh;
-  margin-top: -10px;
-  margin-left: 30px;
-  margin-bottom: 50px;
-  animation-name: floating;
-  animation-duration: 3s;
-  animation-iteration-count: infinite;
-  animation-timing-function: ease-in-out;
-}
-
-@keyframes floating {
-  0% {
-    transform: translate(0, 0px);
-  }
-
-  50% {
-    transform: translate(0, 10px);
-  }
-
-  100% {
-    transform: translate(0, 0px);
-  }
-}
-
-#tassel:hover {
-  filter: drop-shadow(0 0 0.5rem rgb(0 0 0 / 0.4));
-}
-
-#tassel:active {
-  filter: drop-shadow(0 0 0 rgb(0 0 0 / 0.4));
-  animation: stretchjump 1s linear;
-  transform-origin: top center;
-  animation-fill-mode: forwards
-}
-
-@keyframes stretchjump {
-
-  10%,
-  35% {
-    transform: scale(.9, 1.2) translate(0, 0);
-  }
-
-  45%,
-  50% {
-    transform: scale(1.3, .4) translate(0, 0);
-  }
-
-  80% {
-    transform: scale(1) translate(0, 0);
-  }
-}
-
-#TopNav button {
-  background: none;
-  border: none;
-  text-align: center;
-  font-family: inherit;
-  font-size: 18px;
-  text-decoration: none;
-  margin: 5px;
-  padding: 5px 10px;
-  word-spacing: 10px;
-}
-
-.Content {
-  /* Content class covers the entire viewport */
-  height: 100vh;
-  width: 100vw;
-  /* Initialize flexbox */
-  display: flex;
-  flex-direction: row;
-  flex-wrap: wrap;
-  align-items: center;
-}
-
-#resume {
-  position: absolute;
-  width: 90%;
-  height: 90%;
-  z-index: 5;
-}
