+++ src/components/pages/Projects.css
@@ -1,0 +1,3 @@
+.project {
+  display: flex;
+}
